requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
blinker==1.6.2
blis==0.7.9
build==0.10.0
CacheControl==0.12.11
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chromadb==0.3.25
cleo==2.0.1
click==8.1.3
clickhouse-connect==0.5.25
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
confection==0.1.0
crashtest==0.4.1
cryptography==41.0.2
cymem==2.0.7
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
dnspython==2.4.0
duckdb==0.8.0
dulwich==0.21.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
executing==1.2.0
fastapi==0.95.2
filelock==3.12.0
Flask==2.3.2
flatbuffers==23.5.9
frozenlist==1.3.3
gensim==4.3.1
greenlet==2.0.2
h11==0.14.0
halo==0.0.31
hnswlib==0.7.0
html5lib==1.1
httpcore==0.17.3
httptools==0.5.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
installer==0.7.0
ipykernel==6.24.0
ipython==8.14.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.17.3
jupyter_client==8.3.0
jupyter_core==5.3.1
keyring==23.13.1
langchain==0.0.202
langchainplus-sdk==0.0.10
langcodes==3.3.0
lockfile==0.12.2
log-symbols==0.0.14
loguru==0.7.0
lz4==4.3.2
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
ndg-httpsclient==0.5.1
nest-asyncio==1.5.6
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
onnxruntime==1.15.0
openai==0.27.7
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.1
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
pinecone-client==2.2.2
pkginfo==1.9.6
platformdirs==3.5.1
poetry==1.5.0
poetry-core==1.6.0
poetry-plugin-export==1.3.1
posthog==3.0.1
preshed==3.0.8
prompt-toolkit==3.0.39
protobuf==4.23.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
Pygments==2.15.1
pyOpenSSL==23.2.0
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==25.1.0
rapidfuzz==2.15.1
regex==2023.5.5
requests==2.31.0
requests-toolbelt==1.0.0
scipy==1.11.0
shellingham==1.5.0.post1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spinners==0.0.24
SQLAlchemy==2.0.15
srsly==2.4.6
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
termcolor==2.3.0
thinc==8.1.10
tiktoken==0.4.0
tokenizers==0.13.3
tomlkit==0.11.8
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trove-classifiers==2023.5.24
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.22.0
virtualenv==20.23.0
wasabi==1.1.2
watchfiles==0.19.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.3.6
win32-setctime==1.1.0
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
